csv fixes
